requirements.txt
spacy>=3.7.2
streamlit>=1.33.0
pdfminer.six>=20221105
python-docx>=0.8.11
nltk>=3.8.1
scikit-learn>=1.4.2
streamlit-lottie
requests
pandas>=1.3.0